Created metric detail report for each entity, including events and health rules and timeline view Updated detected entity report with links to metrics detail report Created health rule and events report Changed all entity ID ints to be longs Added password encryption for controller password Added all SSL/TLS protocols to accepted ones to accomodate older controllers running with SSLv3 Changed snapshot retrieval batch size to 600 to avoid retrieving unexpected data

+++ DataObjects/RestApi/AppDRESTBackend.cs
@@ -28,10 +28,10 @@ namespace AppDynamics.Dexter.DataObjects
     {
         public int applicationComponentNodeId { get; set; }
         public string exitPointType { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string name{ get; set; }
         public List<AppDRESTBackendProperty> properties { get; set; }
-        public int tierId { get; set; }
+        public long tierId { get; set; }
 
         public override String ToString()
         {
